requirements.txt
pillow==11.1.0
pipreqs==0.5.0